+++ tests/testServer.spec.js
@@ -75,200 +75,10 @@ describe('Configurar a request HTTPS para enviar o endereço IP', () => {
   });
 });
 
-describe.only('Responder o IP do client', () => {
-
-  let browser;
-  let page;
-  const instructions = fs.readFileSync('./instruction.json', 'utf8');
-  const instructionsString = JSON.parse(instructions.toString());
-  let execToken = execTerminal(`./ngrok authtoken ${instructionsString.token}`);
-  let execNgrok = execTerminal('./ngrok http 8080');
-  let execNode = execTerminal('node src/index.js');
-  beforeEach(async () => {
-    browser = await puppeteer.launch({ args: ['--no-sandbox', '--disable-gpu', '--disable-dev-shm-usage', '--window-size=1920,1080'], headless: true });
-    page = await browser.newPage();
-  });
-
-  afterEach(async () => {
-    execToken.kill();
-    execNgrok.kill();
-    execNode.kill();
-  });
-
-  it('Será validado que ao acessar a url sera possível visualizar o ip do client', async () => {
-
-    
-    execToken.stdout.on('data', ()=>{ });
-
-    
-    execNgrok.stdout.on('data', ()=>{ });
-
-    
-    execNode.stdout.on('data', ()=>{ });
-
-    wait(3000);
-    await page.goto(BASE_URL);
-
-    await page.waitForSelector('a[target="_blank"]');
-    const url =  await page.$$eval('a[target="_blank"]', (nodes) => nodes.map((n) => n.innerText));
-    console.log(url)
-    newPage = await browser.newPage();
-
-    console.log(url[1]);
-    newPage.goto(url[1]);
-    console.log(url[1]);
-    await newPage.waitForSelector(dataTestid('ip'));
-
-    const textIp = await newPage.$$eval(dataTestid('ip'), (nodes) => nodes.map((n) => n.innerText));
-    expect(textIp).not.toBeNull();
-  });
-});
-
-describe.only('Responder informações extraídas através do IP do client', () => {
-  let browser;
-  let page;
-  const instructions = fs.readFileSync('./instruction.json', 'utf8');
-  const instructionsString = JSON.parse(instructions.toString());
-  var execToken = execTerminal(`./ngrok authtoken ${instructionsString.token}`);
-  var execNgrok = execTerminal('./ngrok http 8080');
-  var execNode = execTerminal('node src/index.js');
-
-  beforeEach(async () => {
-    browser = await puppeteer.launch({ args: ['--no-sandbox', '--disable-gpu', '--disable-dev-shm-usage', '--window-size=1920,1080'], headless: true });
-    page = await browser.newPage();
-  });
-
-  afterEach(async () => {
-    execToken.kill();
-    execNgrok.kill();
-    execNode.kill();
-  });
-
-  it('Será validado que as informações da localização do cliente serão exibidas na tela', async () => {
 
 
-    execToken.stdout.on('data', ()=>{ });
-  
-    
-    execNgrok.stdout.on('data', ()=>{ });
 
- 
-    execNode.stdout.on('data', ()=>{ });
 
-    wait(3000);
-    await page.goto(BASE_URL);
 
-    await page.waitForSelector('a[target="_blank"]');
-    const url =  await page.$$eval('a[target="_blank"]', (nodes) => nodes.map((n) => n.innerText));
-    console.log(url)
-    newPage = await browser.newPage();
 
-    console.log(url[1]);
-    newPage.goto(url[1]);
-    console.log(url[1]);
-    await newPage.waitForSelector(dataTestid('city'));
-    const textCity = await newPage.$$eval(dataTestid('city'), (nodes) => nodes.map((n) => n.innerText));
-    await newPage.waitForSelector(dataTestid('postal_code'));
-    const textPostaCode = await newPage.$$eval(dataTestid('postal_code'), (nodes) => nodes.map((n) => n.innerText));
-    await newPage.waitForSelector(dataTestid('region'));
-    const textRegion = await newPage.$$eval(dataTestid('region'), (nodes) => nodes.map((n) => n.innerText));
-    await newPage.waitForSelector(dataTestid('country'));
-    const textCountry = await newPage.$$eval(dataTestid('country'), (nodes) => nodes.map((n) => n.innerText));
-    await newPage.waitForSelector(dataTestid('company'));
-    const textCompany = await newPage.$$eval(dataTestid('company'), (nodes) => nodes.map((n) => n.innerText));
 
-    expect(textCity).not.toBeNull();
-    expect(textPostaCode).not.toBeNull();
-    expect(textRegion).not.toBeNull();
-    expect(textCountry).not.toBeNull();
-    expect(textCompany).not.toBeNull();
-  });
-});
-
-describe('Responder dados do dispositivo (client)', () => {
-  it('Será validado se que ao acessar a tela listou os dados do dispositivo', async () => {
-    browser = await puppeteer.launch({ args: ['--no-sandbox', '--disable-gpu', '--disable-dev-shm-usage', '--window-size=1920,1080'], headless: true });
-    page = await browser.newPage();
-
-    const instructions = fs.readFileSync('./instruction.json', 'utf8');
-    const instructionsString = JSON.parse(instructions.toString());
-    var execToken = execTerminal(`./ngrok authtoken ${instructionsString.token}`);
-    execToken.stdout.on('data', ()=>{ });
-
-    var execNgrok = execTerminal('./ngrok http 8080 --log="stdout"');
-    execNgrok.stdout.on('data', ()=>{ });
-
-    var execNode = execTerminal('node src/index.js');
-    execNode.stdout.on('data', ()=>{ });
-
-    wait(2000);
-    await page.goto(BASE_URL);
-    wait(2000);
-
-    await page.waitForSelector('a[target="_blank"]');
-    const url =  await page.$$eval('a[target="_blank"]', (nodes) => nodes.map((n) => n.innerText));
-
-    newPage = await browser.newPage();
-
-    console.log(url[1]);
-    newPage.goto(url[1]);
-    console.log(url[1]);
-    await newPage.waitForSelector(dataTestid('device'));
-    const deviceText = await newPage.$$eval(dataTestid('device'), (nodes) => nodes.map((n) => n.innerText));
-
-    expect(deviceText).not.toBeNull();
-
-    execToken.kill();
-    wait(2000);
-    execNgrok.kill();
-    wait(2000);
-    execNode.kill();
-    wait(2000);
-  });
-});
-
-describe('Responder a request com os resources do Server', () => {
-  it('Validar se acessar o site vai listar as informações do sistema', async () => {
-    browser = await puppeteer.launch({ args: ['--no-sandbox', '--disable-gpu', '--disable-dev-shm-usage', '--window-size=1920,1080'], headless: true });
-    page = await browser.newPage();
-
-    const instructions = fs.readFileSync('./instruction.json', 'utf8');
-    const instructionsString = JSON.parse(instructions.toString());
-    var execToken = execTerminal(`./ngrok authtoken ${instructionsString.token}`);
-    execToken.stdout.on('data', ()=>{ });
-
-    var execNgrok = execTerminal('./ngrok http 8080');
-    execNgrok.stdout.on('data', ()=>{ });
-
-    var execNode = execTerminal('node src/index.js');
-    execNode.stdout.on('data', ()=>{ });
-
-    wait(2000);
-    await page.goto(BASE_URL);
-    wait(2000);
-
-    await page.waitForSelector('a[target="_blank"]');
-    const url =  await page.$$eval('a[target="_blank"]', (nodes) => nodes.map((n) => n.innerText));
-
-    newPage = await browser.newPage();
-
-    newPage.goto(url[1]);
-    await newPage.waitForSelector(dataTestid('arch'));
-    const textArch = await newPage.$$eval(dataTestid('arch'), (nodes) => nodes.map((n) => n.innerText));
-    await newPage.waitForSelector(dataTestid('cpu'));
-    const textCpu = await newPage.$$eval(dataTestid('cpu'), (nodes) => nodes.map((n) => n.innerText));
-    await newPage.waitForSelector(dataTestid('memory'));
-    const textMemory = await newPage.$$eval(dataTestid('memory'), (nodes) => nodes.map((n) => n.innerText));
-
-    expect(textArch).not.toBeNull();
-    expect(textCpu).not.toBeNull();
-    expect(textMemory).not.toBeNull();
-
-    execToken.kill();
-    wait(2000);
-    execNgrok.kill();
-    wait(2000);
-    execNode.kill();
-    wait(2000);
-  });
-});
